# backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: "Smart Receipt API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080/api/v1.0"
    description: "Main backend server (Spring Boot)"
  - url: "http://localhost:8000/api/v1.0"
    description: "AI/ML module server (FastAPI)"
...
paths:
...
  /admin/users:
    get:
      operationId: getAllUsers
      tags:
        - "MainApp"
      summary: "Get all users (admin only)"
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: "List of all registered users"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createUserByAdmin
      tags:
        - "MainApp"
      summary: "Create new user (admin only)"
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserByAdmin'
      responses:
          '201':
            description: "User created successfully"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/UserResponse'
          '400':
            description: "Invalid input data"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '500':
            description: "Internal server error"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{userId}:
    delete:
      operationId: deleteUserByAdmin
      tags:
        - "MainApp"
      summary: "Delete user (admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: "ID of the user to delete"
          schema:
            type: string
            format: uuid
            example: "d4e5f6g7-h8i9-0123-4567-890123fghijk"
      responses:
        '204':
          description: "User deleted successfully"
        '404':
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateUserByAdmin
      tags:
        - "MainApp"
      summary: "Update user details (admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: "ID of the user to update"
          schema:
            type: string
            format: uuid
            example: "d4e5f6g7-h8i9-0123-4567-890123fghijk"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserByAdmin'
      responses:
        '200':
          description: "User updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: "Invalid input data"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum:
        - ROLE_USER
        - ROLE_ADMIN
...
    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          example: "d4e5f6g7-h8i9-0123-4567-890123fghijk"
        email:
          type: string
          format: email
          example: "john.doe@gmail.com"
        role:
          $ref: "#/components/schemas/Role"
    NewUserByAdmin:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
          format: email
          example: "john.doe@gmail.com"
        password:
          type: string
          format: password
          example: "super_secret_password"
        role:
          $ref: "#/components/schemas/Role"
    ErrorResponse:
      required:
        - errorId
        - timestamp
        - status
        - error
        - message
        - path
        - details
      properties:
        errorId:
          type: string
          format: uuid
          example: "f48ac10b-58cc-4372-a567-0e02b2c3d479"
        timestamp:
          type: string
          format: date-time
          example: "2025-11-15T14:30:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Resource not found for this email"
        message:
          type: string
          example: "Resource not found"
        path:
          type: string
          example: "http://localhost:8080/api/v1.0/categories"
        details:
          type: array
          items:
            type: string
          example: ["Resource with provided ID does not exist"]